+++ CHANGELOG.md
@@ -2,10 +2,21 @@
 
 ## Unreleased
 
+### Breaking changes
+
+#### Automatically choose dependency groups strategy when not set
+
+When migrating Poetry or Pipenv dependency groups, `migrate-to-uv` used to always explicitly set all non-optional dependency groups to `default-groups` (e.g., `default-groups = ["dev", "typing"]`).
+
+If no `--dependency-groups-strategy` option is set, it now defaults to using `default-groups = "all"`, unless at least one dependency group is optional, in which case groups are explicitly added to `default-groups`.
+
+If you wish to keep the previous behavior, to always explicitly set all dependency groups, you can set `--dependency-groups-strategy set-default-groups`.
+
 ### Features
 
-* Add `set-default-groups-all` option to `--dependency-groups-strategy` ([#708](https://github.com/mkniewallner/migrate-to-uv/pull/708))
 * Add `--ignore-errors` flag to perform the migration even in case of errors ([#657](https://github.com/mkniewallner/migrate-to-uv/pull/657))
+* Add `set-default-groups-all` option to `--dependency-groups-strategy` ([#708](https://github.com/mkniewallner/migrate-to-uv/pull/708))
+* Automatically choose best dependency groups strategy by default ([#711](https://github.com/mkniewallner/migrate-to-uv/pull/711))
 * [poetry] Bump default `uv_build` bounds to `>=0.10.0,<0.11.0` ([#683](https://github.com/mkniewallner/migrate-to-uv/pull/683))
 
 ### Bug fixes